+++ src/Controller/File/PostElementFileController.php
@@ -16,6 +16,7 @@ use Symfony\Component\Routing\Attribute\Route;
 
 /**
  * @todo add support for etags
+ *
  * @SuppressWarnings("PHPMD.UnusedFormalParameter")
  */
 class PostElementFileController extends AbstractController