+++ src/Controller/User/DeleteTokenController.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Controller\User;
 
 use App\Factory\Exception\Client401UnauthorizedExceptionFactory;
-use App\Factory\Exception\Client404NotFoundExceptionFactory;
 use App\Response\NoContentResponse;
 use App\Security\AuthProvider;
 use App\Service\ElementManager;
@@ -20,7 +19,6 @@ class DeleteTokenController extends AbstractController
         private ElementManager $elementManager,
         private AuthProvider $authProvider,
         private Client401UnauthorizedExceptionFactory $client401UnauthorizedExceptionFactory,
-        private Client404NotFoundExceptionFactory $client404NotFoundExceptionFactory,
     ) {
     }